prevent id update

+++ src/main/java/com/saurs/talktome/controllers/PostController.java
@@ -1,18 +1,24 @@
 package com.saurs.talktome.controllers;
 
-import java.net.URI;
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
-
 import com.saurs.talktome.dtos.PostDTO;
 import com.saurs.talktome.models.Post;
 import com.saurs.talktome.repositories.PostRepository;
+import com.saurs.talktome.repositories.UserRepository;
 import com.saurs.talktome.services.PostService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
+
+import java.net.URI;
+import java.util.List;
+
+import static com.saurs.talktome.utils.ControllerUtils.getAuthenticatedUser;
 
 @RestController
 @RequestMapping(value = "/posts")
@@ -23,6 +29,9 @@ public class PostController {
 
   @Autowired
   private PostRepository repository;
+
+  @Autowired
+  private UserRepository userRepository;
 
   @GetMapping
   public ResponseEntity<List<PostDTO>> getAllPosts(@RequestParam(value = "page", defaultValue = "0") Integer page,
@@ -50,32 +59,40 @@ public class PostController {
   }
 
   @PostMapping("/new")
-  public ResponseEntity<PostDTO> createPost(@RequestBody Post post) {
-    // TODO get authenticated user
+  public ResponseEntity<PostDTO> createPost(@RequestBody Post post, @AuthenticationPrincipal UserDetails activeUser) {
+    Long userId = userRepository.findByUsername(activeUser.getUsername()).get().getId();
     Post newPost = service.addPost(post, userId);
-    URI uri = ServletUriComponentsBuilder
-    .fromCurrentRequest()
-    .path("/{userId}")
-    .buildAndExpand(newPost.getId())
-    .toUri();
-
-    PostDTO postDto = repository.findById(newPost.getId()).map(PostDTO::converter).get();
-
-    return ResponseEntity.created(uri).body(postDto);
+    if (userId.equals(post.getAuthor().getId())) {
+      URI uri = ServletUriComponentsBuilder
+              .fromCurrentRequest()
+              .path("/{userId}")
+              .buildAndExpand(newPost.getId())
+              .toUri();
+      PostDTO postDto = repository.findById(newPost.getId()).map(PostDTO::converter).get();
+      return ResponseEntity.created(uri).body(postDto);
+    }
+    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
   }
 
   @PutMapping("/{id}")
-  public ResponseEntity<PostDTO> updatePost(@RequestBody Post post, @PathVariable Long id) {
-    // TODO get authenticated user and check if it's him
-    Post updatedPost = service.updatePost(post, id);
-    PostDTO postDto = repository.findById(updatedPost.getId()).map(PostDTO::converter).get();
-    return ResponseEntity.ok().body(postDto);
+  public ResponseEntity<PostDTO> updatePost(@RequestBody Post post, @PathVariable Long id, @AuthenticationPrincipal UserDetails activeUser) {
+    Long userId = userRepository.findByUsername(activeUser.getUsername()).get().getId();
+    if (userId.equals(post.getAuthor().getId())) {
+      Post updatedPost = service.updatePost(post, id);
+      PostDTO postDto = repository.findById(updatedPost.getId()).map(PostDTO::converter).get();
+      return ResponseEntity.ok().body(postDto);
+    }
+    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
   }
 
   @DeleteMapping("/{id}")
-  public ResponseEntity<Void> deletePost(@PathVariable Long id) {
-    // TODO get authenticated user and check if it's him
-    service.deletePost(id);
-    return ResponseEntity.noContent().build();
+  public ResponseEntity<Void> deletePost(@PathVariable Long id, @AuthenticationPrincipal UserDetails activeUser) {
+    Long userId = userRepository.findByUsername(activeUser.getUsername()).get().getId();
+    Long postId = repository.findById(id).get().getId();
+    if (userId.equals(postId)) {
+      service.deletePost(id);
+      return ResponseEntity.noContent().build();
+    }
+    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
   }
 }
