+++ src/main/java/com/saurs/talktome/controllers/ReplyController.java
@@ -1,31 +1,37 @@
 package com.saurs.talktome.controllers;
 
-import com.saurs.talktome.dtos.PostDTO;
 import com.saurs.talktome.dtos.ReplyDTO;
-import com.saurs.talktome.models.Post;
 import com.saurs.talktome.models.Reply;
-import com.saurs.talktome.repositories.PostRepository;
 import com.saurs.talktome.repositories.ReplyRepository;
-import com.saurs.talktome.services.PostService;
+import com.saurs.talktome.repositories.UserRepository;
 import com.saurs.talktome.services.ReplyService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import java.net.URI;
+import java.security.Principal;
 import java.util.List;
+
+import static com.saurs.talktome.utils.ControllerUtils.getAuthenticatedUser;
 
 @RestController
 @RequestMapping(value = "/replies")
 public class ReplyController {
-  
   @Autowired
   private ReplyService service;
 
   @Autowired
   private ReplyRepository repository;
+
+  @Autowired
+  private UserRepository userRepository;
 
   @GetMapping
   public ResponseEntity<List<ReplyDTO>> getAllReplies(@RequestParam(value = "page", defaultValue = "0") Integer page,
@@ -62,35 +68,44 @@ public class ReplyController {
   }
 
   @PostMapping("/reply-to/{postId}")
-  public ResponseEntity<ReplyDTO> createReply(@RequestBody Reply reply, @PathVariable Long postId) {
-    // TODO get authenticated user
+  public ResponseEntity<ReplyDTO> createReply(@RequestBody Reply reply, @PathVariable Long postId, Authentication auth, @AuthenticationPrincipal UserDetails activeUser) {
+    Long userId = userRepository.findByUsername(activeUser.getUsername()).get().getId();
     Reply newReply = service.addReply(reply, userId, postId);
-    URI uri = ServletUriComponentsBuilder
-    .fromCurrentRequest()
-    .path("/{userId}")
-    .buildAndExpand(newReply.getId())
-    .toUri();
 
-    ReplyDTO replyDto = repository.findById(newReply.getId()).map(ReplyDTO::converter).get();
+    if (userId.equals(newReply.getAuthor().getId())) {
+      URI uri = ServletUriComponentsBuilder
+              .fromCurrentRequest()
+              .path("/{userId}")
+              .buildAndExpand(newReply.getId())
+              .toUri();
 
-    return ResponseEntity.created(uri).body(replyDto);
+      ReplyDTO replyDto = repository.findById(newReply.getId()).map(ReplyDTO::converter).get();
+
+      return ResponseEntity.created(uri).body(replyDto);
+    }
+    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
   }
 
   @PutMapping("/{id}")
-  public ResponseEntity<ReplyDTO> updateReply(@RequestBody Reply reply, @PathVariable Long id) {
-    // TODO get authenticated user and check if it's him
-
-    Reply updatedReply = service.updateReply(reply, id);
-
-    ReplyDTO replyDto = repository.findById(updatedReply.getId()).map(ReplyDTO::converter).get();
-    return ResponseEntity.ok().body(replyDto);
+  public ResponseEntity<ReplyDTO> updateReply(@RequestBody Reply reply, @PathVariable Long id, @AuthenticationPrincipal UserDetails activeUser) {
+    Long userId = userRepository.findByUsername(activeUser.getUsername()).get().getId();
+    Long replyId = repository.findById(id).get().getId();
+    if (userId.equals(replyId)) {
+      Reply updatedReply = service.updateReply(reply, id);
+      ReplyDTO replyDto = repository.findById(updatedReply.getId()).map(ReplyDTO::converter).get();
+      return ResponseEntity.ok().body(replyDto);
+    }
+    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
   }
 
   @DeleteMapping("/{id}")
-  public ResponseEntity<Void> deleteReply(@PathVariable Long id) {
-    // TODO get authenticated user and check if it's him
-
-    service.deleteReply(id);
-    return ResponseEntity.noContent().build();
+  public ResponseEntity<Void> deleteReply(@PathVariable Long id, @AuthenticationPrincipal UserDetails activeUser) {
+    Long userId = userRepository.findByUsername(activeUser.getUsername()).get().getId();
+    Long replyId = repository.findById(id).get().getId();
+    if (userId.equals(replyId)) {
+      service.deleteReply(id);
+      return ResponseEntity.noContent().build();
+    }
+    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
   }
 }
